refactor: update project data and types to follow the new convention
- Moved project interface to dedicated types file
- Modified projects data to follow the updated naming and export convention
