+++ src/data/index.ts
@@ -1,1 +1,2 @@
+export { PROJECTS } from './projects';
 export { SKILLS } from './skills';
