+++ src/types/index.ts
@@ -1,1 +1,2 @@
 export type { Skill } from './skills';
+export type { Project } from './projects';
