+++ src/types/projects.ts
@@ -1,0 +1,7 @@
+export interface Project {
+  name: string;
+  img: string;
+  demo: string;
+  repo: string;
+  descriptionKey: string;
+}
